bobcat tracked(dynamic arm) *complete overhaul*, The hydraulic mechanism is now physically accurate (*prismatic* joint drives the system instead of the revolute ones) *Visual* model updated to match the static arm.
"Bobcat tracked static arm" center of mass centered so that the sensors are not pitched.
*Sensor locations tweaked* to match the wheeled model.

Bobcat tracked's sensors will never be accurately positions and pitched
due to the *shifting* of the center of mass when moving the arm.

Arm is moved using the Bobcat/arm/* topics input is -1..1 indicating how fast to move the arm.
The plugin is fool proof and includes target limits and collision lash-back protection.

Signed-off-by: yossi2010 <yossicohen2000@gmail.com>